+++ packages/guides/src/Interlink/InventoryLoader.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace phpDocumentor\Guides\Intersphinx;
+namespace phpDocumentor\Guides\Interlink;
 
 use function is_array;
 use function strval;
