+++ packages/guides/src/Interlink/JsonLoader.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace phpDocumentor\Guides\Intersphinx;
+namespace phpDocumentor\Guides\Interlink;
 
 use JsonException;
 use RuntimeException;
